Setup a basic sprite-zero based HUD split
Gosh this is a wasteful CPU spend, but it *works*, so... it'll
do for now. I think the NMI setup is even lag frame safe, so long
as we don't manage to somehow clobber sprite zero with decay, but
we aren't disabling rendering so that shouldn't be a major issue.